requirements.txt
Django==3.1.6
djangorestframework~=3.11.0
django-cors-headers~=3.2.1
djangorestframework-simplejwt~=4.4.0
PyJWT==1.7.1
djoser~=2.1.0
